Introduce LoginPresenter. Handles all login logic
