--- app/src/main/java/com/tristanwiley/chatse/util/ViewUtils.kt
@@ -1,11 +1,0 @@
-package com.tristanwiley.chatse.util
-
-import android.view.View
-
-fun View.show() = apply {
-    visibility = View.VISIBLE
-}
-
-fun View.hide() = apply {
-    visibility = View.GONE
-}
